# Introduce cache for speeding up APIServices lookup