**What type of PR is this?**
/kind cleanup

**What this PR does / why we need it**:
This PR addresses the TODO item above AvailableConditionController#addService

A cache is introduced into AvailableConditionController. In getAPIServicesFor, we check whether the v1.Service reference has been queried before.
If so, return the cached result directly.

```release-note
NONE
```