@liggitt 2019-07-11
given the volume of endpoint updates, I think optimizing getAPIServicesFor is probably worthwhile, but I would recommend something more like https://github.com/kubernetes/kubernetes/pull/79999#discussion_r302735077

this will also need thorough tests to prove that APIService additions/updates/removals result in the proper map of service-namespace/service-name -> apiservice name list