@liggitt 2019-07-11
and since this is being done in service of performance, quantifying the gain with a unit test benchmark is recommended