@tedyu 2019-07-11
@liggitt 
I tried to address your comments.
Please take a look.

I will work on benchmark next.