@tedyu 2019-07-12
/test pull-kubernetes-bazel-test