@tedyu 2019-07-12
For benchmark test, I was looking at endpoints_controller_test.go
But it seems XXEndpoints funcs are not exercised by the test.