@tedyu 2019-07-12
@liggitt 
See if this version gets closer.