@liggitt 2019-07-12
there's still a mixup here. this controller is watching three types of objects: APIService, Service, Endpoints

the **APIService** objects are what should trigger rebuilding the cache, not the **Service** objects