@tedyu 2019-07-12
/test pull-kubernetes-e2e-gce