@tedyu 2019-07-13
I added preliminary benchmark
With cache:
```
pkg: k8s.io/kubernetes/staging/src/k8s.io/kube-aggregator/pkg/controllers/status
BenchmarkBuildCache-12    	  300000	      5364 ns/op	    4048 B/op	      38 allocs/op
```
Without cache:
```
pkg: k8s.io/kubernetes/staging/src/k8s.io/kube-aggregator/pkg/controllers/status
BenchmarkBuildCache-12    	  200000	      5551 ns/op	    4107 B/op	      42 allocs/op
```
I am ready to modify benchmark upon review(s)