@tedyu 2019-07-14
For the following operations as the loop body in the benchmark:
```
		for _, svc := range services {
			c.addService(svc)
		}
		for _, svc := range services {
			c.deleteService(svc)
		}
		c.sync(apiServiceName)
```
With caching:
```
BenchmarkBuildCache-12            200000              5706 ns/op            4059 B/op         39 allocs/op

BenchmarkBuildCache-12            300000              5446 ns/op            4064 B/op         39 allocs/op
```
Without caching:
```
BenchmarkBuildCache-12            200000              6091 ns/op            4187 B/op         47 allocs/op

BenchmarkBuildCache-12            200000              6304 ns/op            4187 B/op         47 allocs/op
```