@tedyu 2019-07-15
The slowest benchmark run with caching exhibited 5% speedup against the fastest run without caching.
The fastest benchmark run with caching exhibited 10% speedup against the fastest run without caching.
In all cases, objection allocation was reduced 15%.

@liggitt 
Does the performance improvement look good ?