@tedyu 2019-07-16
If updateService is involved in the benchmark, the performance improvement is more pronounced.
With caching:
```
BenchmarkBuildCache-12    	  200000	      7400 ns/op	    4203 B/op	      48 allocs/op

BenchmarkBuildCache-12    	  200000	      7240 ns/op	    4203 B/op	      48 allocs/op
```
Without caching:
```
BenchmarkBuildCache-12    	  200000	     10535 ns/op	    4908 B/op	      92 allocs/op

BenchmarkBuildCache-12    	  200000	     10674 ns/op	    4907 B/op	      92 allocs/op
```