@tedyu 2019-07-16
/test pull-kubernetes-integration