@tedyu 2019-07-17
On my Mac, the new benchmark with caching:
```
BenchmarkBuildCache-12    	   50000	     22732 ns/op	      48 B/op	       3 allocs/op

BenchmarkBuildCache-12    	  100000	     23613 ns/op	      48 B/op	       3 allocs/op
```