@liggitt 2019-07-17
benchmark reproducing normal conditions definitely proves the value, will sweep for code style then will be glad to see this merge