@tedyu 2019-07-17
/test pull-kubernetes-integration