@liggitt 2019-07-18
This looks reasonable to me. @deads2k can take a look early next week and it should be good to merge after that