@tedyu 2019-07-25
@deads2k @liggitt 
I have added the comment above rebuildAPIServiceCache func.

Please take another look.

Thanks