@tedyu 2019-07-26
@deads2k @liggitt 
There was bug in my previous change where if castObj or oldCastObj is nil, there would still be panic.

This has been fixed in the latest PR.

Please take a look.