@tedyu 2019-07-29
@deads2k 
Let me know if the latest code addresses your comment.

Thanks