@tedyu 2019-07-30
@deads2k 
Do you have any comment that I should address ?

thanks